pyproject.toml
[project]
name = "ai-image-app"
version = "0.1.0"
description = "AI Image App MVP v0.1 (U1 thin slice)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "celery>=5.3",
  "redis>=5.0",
  "boto3>=1.34",
  "python-json-logger>=2.0",
  "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.10",
  "pytest>=8.0",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "requests>=2.31",
  "types-redis>=4.6",
  "types-boto3>=1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = ["I", "UP", "B"]
exclude = ["alembic", ".venv", "dist", "build"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
